# DH.Curbside.API/App_Data/CurbsideApiSwaggerSpec.json
{  
   "swagger":"2.0",
   "info":{  
      "description":"This is a swagger spec API documentation for Curbside Application",
      "version":"1.0.0",
      "title":"Curbside API"
   },
...
   "paths":{  
...
      "/{tenantId}/case":{  
         "post":{  
            "tags":[  
               "case"
            ],
            "summary":"Add a new case for a patient",
            "operationId":"addPatientCase",
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"path",
                  "name":"tenantId",
                  "description":"organization guid",
                  "required":true,
                  "type":"string"
               },
               {  
                  "in":"body",
                  "name":"body",
                  "required":true,
                  "schema":{  
                     "$ref":"#definitions/PatientCase"
                  }
               }
            ],
            "responses":{  
               "201":{  
                  "description":"Created"
               }
            }
         },
         "get":{  
            "tags":[  
               "case"
            ],
            "summary":"get list of cases, given providerId return only submitted or reviewed",
            "operationId":"getPatientCase",
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"path",
                  "name":"tenantId",
                  "description":"organization guid",
                  "required":true,
                  "type":"string"
               },
               {  
                  "in":"query",
                  "name":"providerId",
                  "required":false,
                  "type":"string"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK",
                  "schema":{  
                     "type":"array",
                     "$ref":"#definitions/PatientCase"
                  }
               }
            }
         }
      },
...
   },
...
   "host":"virtserver.swaggerhub.com",
   "basePath":"/dignity/curbsideapi/1.0.0",
   "schemes":[  
      "https"
   ]
}